requirements.txt
python-multipart==0.0.5
alembic==1.6.5
fastapi==0.75.0
uvicorn==0.17.6
sqlalchemy==1.4.27
pymysql==1.0.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
requests==2.26.0